Tweak the names of equality operators.

+++ slosh/bench_utils/src/lib.rs
@@ -104,7 +104,7 @@ pub fn run_float_script(n: usize, m: f32, expected: f64) {
     let last = run_reader(&mut reader);
     match last {
         Ok(Value::Float(f)) => {
-            assert_eq!(expected, f64::from(f));
+            assert_eq!(expected as f32, f32::from(f));
         }
         _ => {
             panic!("Not a float");
